analytics/requirements.txt
# Pirata Pizzaria - Analytics & ML
# Python 3.10+

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# ML & Forecasting
scikit-learn>=1.3.0
prophet>=1.1.4
statsmodels>=0.14.0

# Weather API
requests>=2.31.0
openmeteo-requests>=1.2.0
requests-cache>=1.1.0

# Visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# API (FastAPI microservice)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utils
python-dotenv>=1.0.0
schedule>=1.2.0

# Jupyter (for analysis)
jupyter>=1.0.0
ipykernel>=6.25.0

